Update asset links
