+++ navigation/MenuButton/styles.js
@@ -1,0 +1,12 @@
+import { StyleSheet } from "react-native";
+
+const styles = StyleSheet.create({
+  menuIcon: {
+    zIndex: 9,
+    position: "absolute",
+    top: 40,
+    left: 20
+  }
+});
+
+export default styles;
